requirements.txt
speechrecognition-3.14.1
pyaudio-0.2.14
torch-2.6.0
sympy-1.13.1
torchaudio-2.6.0
transformers-4.49.0
librosa-0.10.2
soundfile-0.13.1
audioread-3.0.1
cffi-1.17.1
decorator-5.1.1
huggingface_hub-0.28.1
fsspec-2025.2.0
joblib-1.4.2
lazy_loader-0.4
msgpack-1.1.0
numba-0.61.0
numpy-2.1.3
pooch-1.8.2
PyYAML-6.0.2
regex-2024.11.6
safetensors-0.5.2
scikit_learn-1.6.1
scipy-1.15.2
tokenizers-0.21.0
